added relations between articles and articles

+++ src/articles/articles.service.ts
@@ -19,22 +19,22 @@ export class ArticlesService {
         return this.articleRepository.getArticles();
     }
 
-    async createArticle(articleDto: ArticleDto): Promise<Article> {
+    async createArticle(articleDto: ArticleDto, user): Promise<Article> {
        const category = await Category.findOne(articleDto.categoryId);
 
        if(!category) {
             throw new NotFoundException();
        }
 
-       return await this.articleRepository.createArticle(articleDto, category);
+       return await this.articleRepository.createArticle(articleDto, category, user);
     }
 
-    async updateArticle(id: number, articleDto: ArticleDto): Promise<Article> {
-        return await this.articleRepository.updateArticle(id, articleDto);
+    async updateArticle(id: number, articleDto: ArticleDto, user): Promise<Article> {
+        return await this.articleRepository.updateArticle(id, articleDto, user);
     }
 
-    async deleteArticle(id: number) {
-        const result = await this.articleRepository.delete({ id });
+    async deleteArticle(id: number, user) {
+        const result = await this.articleRepository.delete({ id: id, userId: user.id });
 
         if(result.affected === 0) {
             throw new NotFoundException();
